+++ zarrs/src/array/chunk_grid/regular.rs
@@ -9,6 +9,7 @@ use zarrs_registry::chunk_grid::REGULAR;
 
 use crate::{
     array::{chunk_grid::ChunkGridPlugin, ArrayIndices, ArrayShape, ChunkShape},
+    array_subset::{ArraySubset, IncompatibleDimensionalityError},
     metadata::v3::MetadataV3,
     plugin::{PluginCreateError, PluginMetadataInvalidError},
 };
@@ -100,12 +101,91 @@ impl RegularChunkGrid {
 
     /// Return the chunk shape as an [`ArrayShape`] ([`Vec<u64>`]).
     #[must_use]
-    pub fn chunk_shape_u64(&self) -> Vec<u64> {
+    pub fn chunk_shape_u64(&self) -> ArrayShape {
         self.chunk_shape
             .iter()
             .copied()
             .map(NonZeroU64::get)
-            .collect::<Vec<_>>()
+            .collect::<ArrayShape>()
+    }
+
+    /// Return the origin of the chunk at `chunk_indices`.
+    ///
+    /// # Errors
+    /// Returns [`IncompatibleDimensionalityError`] if `chunk_indices` do not match the dimensionality of the chunk grid.
+    pub fn chunk_origin(
+        &self,
+        chunk_indices: &[u64],
+    ) -> Result<ArrayIndices, IncompatibleDimensionalityError> {
+        if chunk_indices.len() == self.dimensionality() {
+            Ok(std::iter::zip(chunk_indices, self.chunk_shape.as_slice())
+                .map(|(i, s)| i * s.get())
+                .collect())
+        } else {
+            Err(IncompatibleDimensionalityError::new(
+                chunk_indices.len(),
+                self.dimensionality(),
+            ))
+        }
+    }
+
+    /// The indices of a chunk which has the element at `array_indices`.
+    ///
+    /// # Errors
+    /// Returns [`IncompatibleDimensionalityError`] if `array_indices` do not match the dimensionality of the chunk grid.
+    fn chunk_indices(
+        &self,
+        array_indices: &[u64],
+    ) -> Result<ArrayIndices, IncompatibleDimensionalityError> {
+        if array_indices.len() == self.dimensionality() {
+            Ok(std::iter::zip(array_indices, self.chunk_shape.as_slice())
+                .map(|(i, s)| i / s.get())
+                .collect())
+        } else {
+            Err(IncompatibleDimensionalityError::new(
+                array_indices.len(),
+                self.dimensionality(),
+            ))
+        }
+    }
+
+    /// Return the [`ArraySubset`] of the chunk at `chunk_indices`.
+    ///
+    /// # Errors
+    /// Returns [`IncompatibleDimensionalityError`] if `chunk_indices` do not match the dimensionality of the chunk grid.
+    pub fn subset(
+        &self,
+        chunk_indices: &[u64],
+    ) -> Result<ArraySubset, IncompatibleDimensionalityError> {
+        let chunk_origin = self.chunk_origin(chunk_indices)?;
+        let ranges = chunk_origin
+            .iter()
+            .zip(self.chunk_shape.iter())
+            .map(|(&o, &s)| o..(o + s.get()));
+        Ok(ArraySubset::from(ranges))
+    }
+
+    /// Return an array subset indicating the chunks intersecting `array_subset`.
+    ///
+    /// # Errors
+    /// Returns [`IncompatibleDimensionalityError`] if `array_subset` does not match the dimensionality of the chunk grid.
+    pub fn chunks_in_array_subset(
+        &self,
+        array_subset: &ArraySubset,
+    ) -> Result<ArraySubset, IncompatibleDimensionalityError> {
+        match array_subset.end_inc() {
+            Some(end) => {
+                let chunks_start = self.chunk_indices(array_subset.start())?;
+                let chunks_end = self.chunk_indices(&end)?;
+                // .unwrap_or_else(|| self.grid_shape());
+
+                let shape = std::iter::zip(&chunks_start, chunks_end)
+                    .map(|(&s, e)| e.saturating_sub(s) + 1)
+                    .collect();
+                Ok(ArraySubset::new_with_start_shape(chunks_start, shape)?)
+            }
+            None => Ok(ArraySubset::new_empty(self.dimensionality())),
+        }
     }
 }
 
@@ -139,31 +219,15 @@ impl ChunkGridTraits for RegularChunkGrid {
     /// The chunk shape as an [`ArrayShape`] ([`Vec<u64>`]). Fixed for a `regular` grid.
     unsafe fn chunk_shape_u64_unchecked(&self, chunk_indices: &[u64]) -> Option<ArrayShape> {
         debug_assert_eq!(self.dimensionality(), chunk_indices.len());
-        Some(
-            self.chunk_shape
-                .iter()
-                .copied()
-                .map(NonZeroU64::get)
-                .collect::<ArrayShape>(),
-        )
+        Some(self.chunk_shape_u64())
     }
 
     unsafe fn chunk_origin_unchecked(&self, chunk_indices: &[u64]) -> Option<ArrayIndices> {
-        debug_assert_eq!(self.dimensionality(), chunk_indices.len());
-        Some(
-            std::iter::zip(chunk_indices, self.chunk_shape.as_slice())
-                .map(|(i, s)| i * s.get())
-                .collect(),
-        )
+        Some(self.chunk_origin(chunk_indices).unwrap())
     }
 
     unsafe fn chunk_indices_unchecked(&self, array_indices: &[u64]) -> Option<ArrayIndices> {
-        debug_assert_eq!(self.dimensionality(), array_indices.len());
-        Some(
-            std::iter::zip(array_indices, self.chunk_shape.as_slice())
-                .map(|(i, s)| i / s.get())
-                .collect(),
-        )
+        Some(self.chunk_indices(array_indices).unwrap())
     }
 
     unsafe fn chunk_element_indices_unchecked(
@@ -176,6 +240,13 @@ impl ChunkGridTraits for RegularChunkGrid {
                 .map(|(i, s)| i % s.get())
                 .collect(),
         )
+    }
+
+    fn chunks_in_array_subset(
+        &self,
+        array_subset: &ArraySubset,
+    ) -> Result<Option<ArraySubset>, IncompatibleDimensionalityError> {
+        Ok(Some(self.chunks_in_array_subset(array_subset)?))
     }
 }
 
@@ -227,17 +298,14 @@ mod tests {
             let chunk_grid =
                 RegularChunkGrid::new(array_shape.clone(), chunk_shape.clone()).unwrap();
             assert_eq!(chunk_grid.dimensionality(), 3);
-            assert_eq!(
-                chunk_grid.chunk_origin(&[1, 1, 1]).unwrap(),
-                Some(vec![1, 2, 3])
-            );
+            assert_eq!(chunk_grid.chunk_origin(&[1, 1, 1]).unwrap(), vec![1, 2, 3]);
             assert_eq!(chunk_grid.chunk_shape(), chunk_shape.as_slice());
             let chunk_grid_shape = chunk_grid.grid_shape();
             assert_eq!(chunk_grid_shape, &[5, 4, 18]);
             let array_index: ArrayIndices = vec![3, 5, 50];
             assert_eq!(
                 chunk_grid.chunk_indices(&array_index).unwrap(),
-                Some(vec![3, 2, 16])
+                vec![3, 2, 16]
             );
             assert_eq!(
                 chunk_grid.chunk_element_indices(&array_index).unwrap(),
@@ -274,14 +342,14 @@ mod tests {
         let array_indices: ArrayIndices = vec![3, 5, 53];
         assert_eq!(
             chunk_grid.chunk_indices(&array_indices).unwrap(),
-            Some(vec![3, 2, 17])
+            vec![3, 2, 17]
         );
 
         let chunk_indices: ArrayShape = vec![6, 1, 1];
         assert!(!chunk_grid.chunk_indices_inbounds(&chunk_indices));
         assert_eq!(
             chunk_grid.chunk_origin(&chunk_indices).unwrap(),
-            Some(vec![6, 2, 3])
+            vec![6, 2, 3]
         );
     }
 
@@ -292,12 +360,14 @@ mod tests {
         let chunk_grid = RegularChunkGrid::new(array_shape, chunk_shape).unwrap();
 
         let array_indices: ArrayIndices = vec![3, 5, 1000];
-        assert!(chunk_grid.chunk_indices(&array_indices).unwrap().is_some());
+        assert_eq!(
+            chunk_grid.chunk_indices(&array_indices).unwrap(),
+            vec![3, 2, 333]
+        );
 
         assert_eq!(chunk_grid.grid_shape(), &[5, 4, 0]);
 
         let chunk_indices: ArrayShape = vec![3, 1, 1000];
         assert!(chunk_grid.chunk_indices_inbounds(&chunk_indices));
-        assert!(chunk_grid.chunk_origin(&chunk_indices).unwrap().is_some());
     }
 }